Move launch desktop logic to the button component
Also start yanking the user's input name when launching the session
